requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
cloudinary
fpdf2
flask-cors
